Create README.md
This research paper delves into the realm of machine learning-driven fraud detection, focusing on the deployment of Amazon SageMaker as a comprehensive solution. The primary objective of this study is to develop a fraud detection framework that possesses adaptability, self-improvement, and scalability, all crucial aspects for businesses operating in the online sphere. The paper showcases the integration of SageMaker into the fraud detection pipeline, highlighting its role in fostering an iterative approach to model enhancement and adaptation. By harnessing SageMaker's capabilities, businesses can proactively combat fraud, swiftly respond to evolving tactics, and scale their defenses according to the demands of their online operations.